mv test.py to app.py
